Return Simulation from multivariate normal distribution

Simulate self-independent returns using Choleski Decomposition where correlation depends on the order of assets from multvariate normal distribution

sim_simple(dis_par = list(mu = 0.05, vol = 0.02, df = 2),
  distribution = c("norm", "t"), Tn = 10, N = 5, M = 10000,
  varcov = NULL, par = 0.2, rho_do = NULL, dependent = c(NA,
  "AR1"), den_par = list(AR1 = list(beta_0 = 0.05, beta_1 = 0.5)),
  return_varcov = FALSE, plus_one = TRUE)

Arguments

Tn
Number of periods, excluding time 0.

N
Number of assets

M
Number of realization

par
The parameter in the default correlation function. See Details.

rho_do
The function of correlation between assets. See below for the default function.

dependent
If the following simulated series are added on the basis of existing ones.

mu
Either a scalar or a vector with length N contains mean of the assets returns

vol
Either a scalar or a vector with length N contains volatilities of the assets returns

varcor
The variance covariance matrix of the assets. Diagonal elements must equal to vol squared. If supplied, arguments par and rho_do will be ignored. If NULL, will be calculated using the correlation function and vol.

Value

List of returns

Details

The correlation function determines the correlation between asset by the difference between the index of the assets. The default function is rho = exp(-par * |distance|). User defined correlation function need to have two arguments i and j to indicate the position of assets